sign_tx: Strengthen liquid commitment checks against the tx being signed
This change ensures that the transaction and any commitment values
match, if an asset/value commitment pair are given.

Commitments given as part of tx input data must have these fields, while
they are optional (and should be discouraged) for tx outputs.

We move the validation checking which was spread over several unrelated
areas into the function fetching the commitment data. This ensures that
when fetching it, it is either fully validated from that point on or an
error is returned.

Following this we remove the commitment fields from the commitment_t struct
since they are now redundant for the remainder of sign_tx processing.

From experience debugging tests where the commitments mismatched, the
old error messages were not helpful. Update these to include 'trusted
commitment' in all messages to make it clearer when a mismatch is
detected.

Note This commit is deliberately minimal to make the core logic changes
easier to review.