WIP: libjade: enable building the Jade firmware as a native library
Implements an in-process instance of the Jade firmware. This can be
thought of as an emulated or virtual Jade device that (initially) runs
in the address space of the users application.

The Jade firmware code runs in a separate thread and receives/replies to
messages using the standard CBOR message structure. An API is provided to
start/stop and send/receive messages to the virtual device.

This can be used to locally test and debug Jade during development.
Since the resulting firmware can be linked and run in-process, it is
trivial to run under a debugger and step through the firmware code.

This initial implementation is highly experimental and incomplete, and
should in no circumstances be used beyond development and testing.