DONTMERGE: attempt to get debug_scan_qr working
It turns out whatever miniz is producing, its not compatible with
compress/uncompress from zlib with the flags we are using. It seems we
don't include any header bytes in the compressed data and so we'd probably
need to prepend a fake header to get this to work.

This commit doesn't enable the debug_scan_qr tests so shouldn't fail the
build.